# openapi.json
{"openapi":"3.0.1","info":{"title":"HR LMS API","description":"HR/Candidate test system API","contact":{"name":"HR LMS"},"version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"basicAuth":[]}],"paths":{"/api/hr/tests/{id}":{"put":{"tags":["hr-controller"],"operationId":"updateTest","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTestRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}},"delete":{"tags":["hr-controller"],"operationId":"deleteTest","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/api/hr/questions/{questionId}":{"put":{"tags":["hr-controller"],"operationId":"updateQuestion","parameters":[{"name":"questionId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateQuestionRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}},"delete":{"tags":["hr-controller"],"operationId":"deleteQuestion","parameters":[{"name":"questionId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/api/hr/candidates/{candidateId}":{"put":{"tags":["hr-controller"],"operationId":"updateCandidate","parameters":[{"name":"candidateId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCandidateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}},"delete":{"tags":["hr-controller"],"operationId":"deleteCandidate","parameters":[{"name":"candidateId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/api/hr/tests":{"get":{"tags":["hr-controller"],"operationId":"listTests","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}},"post":{"tags":["hr-controller"],"operationId":"createTest","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTestRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/api/hr/candidates":{"get":{"tags":["hr-controller"],"operationId":"listCandidates","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}},"post":{"tags":["hr-controller"],"operationId":"createCandidate","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCandidateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/api/candidate/tests/start":{"post":{"tags":["candidate-controller"],"operationId":"start","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StartTestRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/api/candidate/auth/login":{"post":{"tags":["candidate-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/api/candidate/attempts/{attemptId}/submit":{"post":{"tags":["candidate-controller"],"operationId":"submit","parameters":[{"name":"attemptId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubmitAttemptRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/api/candidate/{candidateId}/tests":{"get":{"tags":["candidate-controller"],"operationId":"listRandomTests","parameters":[{"name":"candidateId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}}},"components":{"schemas":{"UpdateTestRequest":{"type":"object","properties":{"title":{"type":"string"},"profession":{"type":"string"},"questionText":{"type":"string"},"active":{"type":"boolean"}}},"ApiResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"type":"object"}}},"OptionRequest":{"required":["correct","text"],"type":"object","properties":{"text":{"type":"string"},"correct":{"type":"boolean"}}},"UpdateQuestionRequest":{"type":"object","properties":{"text":{"type":"string"},"options":{"type":"array","items":{"$ref":"#/components/schemas/OptionRequest"}}}},"UpdateCandidateRequest":{"type":"object","properties":{"fullName":{"type":"string"},"profession":{"type":"string"},"password":{"type":"string"},"active":{"type":"boolean"}}},"CreateTestRequest":{"required":["options","profession","questionText","title"],"type":"object","properties":{"title":{"type":"string"},"profession":{"type":"string"},"questionText":{"type":"string"},"options":{"maxItems":2147483647,"minItems":2,"type":"array","items":{"$ref":"#/components/schemas/OptionRequest"}},"active":{"type":"boolean"}}},"CreateCandidateRequest":{"required":["fullName","login","password","profession"],"type":"object","properties":{"fullName":{"type":"string"},"profession":{"type":"string"},"login":{"type":"string"},"password":{"type":"string"},"active":{"type":"boolean"}}},"StartTestRequest":{"required":["candidateId"],"type":"object","properties":{"candidateId":{"type":"integer","format":"int64"}}},"LoginRequest":{"required":["login","password"],"type":"object","properties":{"login":{"type":"string"},"password":{"type":"string"}}},"AnswerRequest":{"required":["questionId"],"type":"object","properties":{"questionId":{"type":"integer","format":"int64"},"selectedOptionId":{"type":"integer","format":"int64"}}},"SubmitAttemptRequest":{"required":["answers","candidateId"],"type":"object","properties":{"candidateId":{"type":"integer","format":"int64"},"answers":{"type":"array","items":{"$ref":"#/components/schemas/AnswerRequest"}}}}},"securitySchemes":{"basicAuth":{"type":"http","scheme":"basic"}}}}